refactor: add toolbox module
This will make it easier to extend (e.g. with diff support).
