+++ pyproject.toml
@@ -97,6 +97,7 @@ ignore = ["E203", "E501", "E704", "W503"]
 [tool.isort]
 profile = "black"
 force_sort_within_sections = true
+lines_after_imports = 2
 
 [tool.mypy]
 disable_error_code = "import-untyped"
