+++ src/git_draft/__init__.py
@@ -2,6 +2,7 @@ import logging
 
 from .bots import Action, Bot, Toolbox
 
+
 __all__ = [
     "Action",
     "Bot",
