+++ src/git_draft/bots/__init__.py
@@ -8,7 +8,9 @@ import os
 import sys
 
 from ..common import BotConfig, reindent
-from .common import Action, Bot, Goal, Operation, OperationHook, Toolbox
+from ..toolbox import Operation, OperationHook, Toolbox
+from .common import Action, Bot, Goal
+
 
 __all__ = [
     "Action",
