+++ src/git_draft/editor.py
@@ -6,6 +6,7 @@ import subprocess
 import sys
 import tempfile
 
+
 _default_editors = ["vim", "emacs", "nano"]
