+++ tests/git_draft/prompt_test.py
@@ -1,12 +1,14 @@
 import pytest
 
 import git_draft.prompt as sut
+from git_draft.toolbox import StagingToolbox
 
 
 class TestPromptRenderer:
     @pytest.fixture(autouse=True)
     def setup(self, repo) -> None:
-        self._renderer = sut.PromptRenderer.for_repo(repo)
+        toolbox = StagingToolbox(repo)
+        self._renderer = sut.PromptRenderer.for_toolbox(toolbox)
 
     def test_ok(self) -> None:
         prompt = sut.TemplatedPrompt.parse("add-test", "symbol=foo")
